@@ -4,9 +4,12 @@
   "homepage": "https://mocon.github.io/ten-ideas-a-day/",
   "private": true,
   "dependencies": {
+    "firebase": "4.0.0",
     "milligram": "1.3.0",
-    "react": "15.4.2",
-    "react-dom": "15.4.2"
+    "prop-types": "15.5.10",
+    "react": "15.5.4",
+    "react-dom": "15.4.2",
+    "reactfire": "1.0.0"
   },
   "devDependencies": {
     "gh-pages": "1.0.0",
